pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "chromadb>=1.0.20",
    "google-genai>=1.32.0",
    "markdown>=3.8.2",
    "sift-stack-py>=0.8.5",
    "streamlit>=1.49.1",
]
